src/web/requirements.txt
Django==2.2
psycopg2>=2.7
pandas>=0.20.3
django_elasticsearch_dsl
django-widget-tweaks